backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.13.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.118.2
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
narwhals==2.7.0
numpy==2.2.6
openai==2.2.0
packaging==25.0
pandas==2.3.3
passlib[bcrypt]==1.7.4
pillow==11.3.0
pluggy==1.6.0
propcache==0.4.1
protobuf==6.32.1
pyarrow==21.0.0
pydantic==2.12.0
pydantic-settings==2.4.0
pydantic_core==2.41.1
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
redis==5.0.6
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.48.0
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tomli==2.2.1
tornado==6.5.2
tqdm==4.67.1
twilio==9.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
yarl==1.22.0
